removed a compiler warning
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@2116 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ visualization/src/histogram_visualizer.cpp
@@ -354,7 +354,7 @@ pcl::visualization::PCLHistogramVisualizer::addFeatureHistogram (
     const int index,
     const std::string &id, int win_width, int win_height)
 {
-  if (index < 0 || index >= cloud.width * cloud.height)
+  if (index < 0 || index >= (int)(cloud.width * cloud.height))
   {
     PCL_ERROR ("[addFeatureHistogram] Invalid point index (%d) given!\n", index);
     return (false);
